make gcc@pc happy
git-svn-id: http://svn.open64.net/svnroot/open64@1247 ead328b1-e903-0410-a19d-eacee4b031f6

+++ trunk/osprey/be/opt/opt_lmv_helper.cxx
@@ -1451,7 +1451,7 @@ MEM_ACCESS::Print (FILE* f) {
   fprintf (f, "ofst:");
   _ofst.Print (f);
   
-  fprintf (f, ",sz:%d,fld:%d,", _size, _field_id);
+  fprintf (f, ",sz:%d,fld:%d,", (INT)_size, _field_id);
   if (_obj_ty != (TY_IDX)0) {
     fprintf (f, "Ty:%s,", TY_name(_obj_ty));
   }
